+++ app/src/server.py
@@ -195,6 +195,13 @@ class Server():
         else:
             self.trace_level = self.log_level
 
+        if self.trace_level:
+            loggers = ['msg', 'tracer']
+            for logger_name in loggers:
+                logging.getLogger(logger_name).setLevel(self.trace_level)
+        else:
+            self.trace_level = self.log_level
+
     def build_config(self):
         """
         Loads configuration from multiple sources in priority order: